Move the Franka driver tests into their own file
They landed in `test_ik.py`, whose subject is the IK solvers, because nothing in the
repo covered the arm driver. `test_franka.py` mirrors `franka.py` the way `test_ik.py`
mirrors `ik.py`.

Ticket: Positronic-Robotics/internal#154 #open
